% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption.R
\name{assumptions}
\alias{assumptions}
\title{ANOVA assumptions}
\usage{
assumptions(fit)
}
\arguments{
\item{fit}{an object of class 'oneway'.}
}
\value{
two data frames, Q-Q plot and a boxplot
}
\description{
assumptions() assesses ANOVA assumptions of normality of residuals,
and equality of group variances on the outcome variable
}
\examples{
fit <- oneway(mpg~cyl, mtcars)
plot(fit)

}
