% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars.R
\docType{data}
\name{cars}
\alias{cars}
\title{: Car class and highway mileage}
\source{
The data are a subset of code \code{mpg} dataset in the 
\code{\link{ggplot2}{mpg}} dataset.
}
\usage{
cars
}
\description{
A data frame contianing car class and highway miles per gallon
for 38 popular car models between 1999 and 2008
}
\keyword{datasets}
