% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\docType{data}
\name{coagulation}
\alias{coagulation}
\title{Coagulation Information.
A dataset containing the diet groups and the corresponding coag response.}
\format{A data frame with 24 rows and 2 variables:
\describe{
  \item{coag}{num, measured numbers}
  \item{diet}{factor, 4 levels "A", "B", "C", "D"}
  ...
}}
\usage{
coagulation
}
\description{
Coagulation Information.
A dataset containing the diet groups and the corresponding coag response.
}
\keyword{datasets}

