% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{describe}
\alias{describe}
\title{Descriptive statistics}
\usage{
describe(data, formula, digits = 2)
}
\arguments{
\item{data}{A data frame or a list.}

\item{formula}{A formula with a dependent variable (DV) and an independent variable (IV).
For example: \code{DV ~ IV}.}

\item{digits}{Integer (default: 2). Rounding digits.}
}
\value{
A data frame with 13 columns:
\describe{
\item{GROUP}{Group name.}
\item{CLD}{Compact letter display for multiple comparisons. This column is
returned as an empty character vector and is intended to be filled by
post hoc comparison functions. See \code{oneway::compact_letter_display}.}
\item{N}{Sample size.}
\item{AVG}{Arithmetic mean.}
\item{SD}{Sample standard deviation.}
\item{MED}{Median.}
\item{MIN}{Minimum observed value.}
\item{MAX}{Maximum observed value.}
\item{CI (95\%)}{Two-sided 95\% confidence interval for the population mean.}
\item{SKEW (= 0)}{Normal distribution has a skewness of 0. See \code{normality::skewness}.}
\item{KURT (= 3)}{Normal distribution has a kurtosis of 3. See \code{normality::kurtosis}.}
\item{normality}{Is the data normally distributed? See \code{normality::is_normal}.}
\item{n_outliers}{Number of possible outliers. See \code{outlying::Grubbs_test}.}
}
}
\description{
Compute and summarize descriptive statistics for one or more groups.
}
\examples{
y1 <- c(stats::rnorm(20), 7)
y2 <- c(stats::rnorm(22), -7, 9)
describe(list("apple" = y1, "banana" = y2))
}
