% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{lsmeans.oneway}
\alias{lsmeans.oneway}
\title{Fisher LSD test}
\usage{
lsmeans.oneway(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object from oneway function call}

\item{conf.level}{A number of confidence level}
}
\value{
a list of the mean difference and Fisher LSD results
}
\description{
Fisher LSD test
}

