% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{lsmeans}
\alias{lsmeans}
\title{Perform Fisher's LSD}
\usage{
lsmeans(object)
}
\arguments{
\item{object}{A \code{oneway} object}
}
\description{
Test pairwise differences using Fisher's LSD proceduce on a
\code{oneway} object
}

