\name{oneway-package}
\alias{oneway-package}
\alias{oneway}
\docType{package}
\title{
\packageTitle{oneway}
}
\description{
\packageDescription{oneway}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{oneway}
\packageIndices{oneway}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{oneway}

Maintainer: \packageMaintainer{oneway}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{


}
\examples{
#y<-oneway(group)
#summary.oneway(y)
}
