% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway.factor}
\alias{oneway.factor}
\title{Oneway Anova Table Generating in a General Way}
\usage{
\method{oneway}{factor}(x, y, ...)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A factor object}
}
\value{
calling oneway function and return the same result as oneway
}
\description{
Oneway Anova Table Generating in a General Way
}

