% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway.formula}
\alias{oneway.formula}
\title{Oneway Anova Table Generating for R models}
\usage{
\method{oneway}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{An oject for the model.frame}

\item{data}{A list}
}
\value{
calling oneway.factor and return the same result as oneway
}
\description{
Oneway Anova Table Generating for R models
}

