% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{Oneway Anova Table Generating}
\usage{
oneway(z, ...)
}
\arguments{
\item{z}{A data frame or a list to be analyzed}
}
\value{
The list of the anova table including the basic information
}
\description{
Oneway Anova Table Generating
}
\examples{
oneway(list(c(3, 5, 3, 4, 6, 3), c("A", "A", "B", "A", "B", "B")))
}

