% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plasma_etching}
\alias{plasma_etching}
\title{Plasma Etching Experiment}
\format{
A data frame with 20 rows and 2 columns:
\describe{
\item{power}{A factor identifying the power (W) (160(20)220).}
\item{etch_rate}{A numeric response variable.}
}
}
\usage{
plasma_etching
}
\description{
Etch Rate Data (in Å/min) from the Plasma Etching Experiment.
}
\references{
Montgomery, D. C. (2017).
Design and analysis of experiments (Ninth edition).
John Wiley & Sons, Inc.
}
\keyword{datasets}
