% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{posthoc}
\alias{posthoc}
\title{Post-hoc comparisons}
\usage{
posthoc(model)
}
\arguments{
\item{model}{an object of class \code{oneway}}
}
\value{
a data frame
}
\description{
Create a function the tests for group differences, two groups at a time using post-hoc comparisons.
}
\examples{
fit <- oneway(mpg ~ cyl, mtcars)
posthoc(fit)
}
