% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{print.oneway}
\alias{print.oneway}
\title{Print Function to print out the information of oneway anova table}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{A list from oneway function}
}
\value{
the necessory information for a anova table
}
\description{
Print Function to print out the information of oneway anova table
}

