% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{print.summary.oneway}
\alias{print.summary.oneway}
\title{The print functional to print anova table}
\usage{
\method{print}{summary.oneway}(x, ...)
}
\arguments{
\item{x}{An object from oneway return}
}
\value{
an anova table containing the basic information
}
\description{
The print functional to print anova table
}

