% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Oneway ANOVA Report}
\usage{
report(formula, data, browse = TRUE)
}
\arguments{
\item{formula}{an object of class formula, relating the
dependent variable to the grouping variable.}

\item{data}{a data frame containing the variables in the model.}

\item{browse}{logical. If \code{FALSE} display report in Rstudio
instead of browser.}
}
\value{
path to HTML report
}
\description{
Create an HTML ANOVA report
}
\examples{
report(mpg ~ cyl, mtcars)
}
