% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{summary.oneway}
\alias{summary.oneway}
\title{The summary method to create a summary object}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{An object from oneway return}
}
\value{
a table like list includng the anova table information
}
\description{
The summary method to create a summary object
}

