% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tukey_Kramer_test.R
\name{Tukey_Kramer_test}
\alias{Tukey_Kramer_test}
\title{Tukey-Kramer test}
\usage{
Tukey_Kramer_test(
  data,
  formula = NULL,
  alpha = 0.05,
  rounding = 4,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{rounding}{Integer (default: 4). Rounding digits.}

\item{verbose}{Logical (default: TRUE). Show warnings and messages.}
}
\value{
A list with 4 elements:
\describe{
\item{method}{Statistical procedures that were conducted.}
\item{data}{The input data and possibly other transformed data.}
\item{pre_hoc}{\emph{a priori} test result.}
\item{post_hoc}{Post-hoc test result.}
\item{summary}{Descriptive statistics.}
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\examples{
out <- Tukey_Kramer_test(O_O_X, val ~ grp)

}
\references{
Howell, D. C. (2013).
Statistical Methods for Psychology (8th ed.).
Cengage Learning. Chapter 12: Multiple comparisons among treatment means,
Section 12.7: Tukey's test, pg. 394.

Zar, J. H. (2014).
Biostatistical analysis (5th edition).
Pearson. Chapter 11: Multiple comparisons, pg. 244-245.
}
