% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars.R
\docType{data}
\name{cars}
\alias{cars}
\title{Car Class and Highway Mileage}
\format{A data frame with 2034 rows and 3 variables
The variables are as follows:
\describe{
  \item{hwy}{highway miles per gallon}
  \item{class}{car type}
  \item{year}{year of manufacture}
}}
\source{
The data are a subset of \code{mpg} dataset in the
\code{\link{ggplot2}{mpg}} dataset
}
\usage{
cars
}
\description{
\code A dtaframe containing car class and highway miles per gallon for 38 popular
car models between 1999 and 2008
}
\keyword{datasets}
