% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars.R
\docType{data}
\name{cars}
\alias{cars}
\title{Car class and highway mileage}
\format{
A data frame with 234 rows and 3 variables:
\describe{
\item{\code{hwy}}{integer. highway miles per gallon.}
\item{\code{class}}{character. car type.}
\item{\code{year}}{integer. year of manufacture.}
}
}
\usage{
cars
}
\description{
A dataset containing car class and highway miles
per gallon for 38 popular car models between 1999 and 2008
}
\keyword{datasets}
