% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_coagulation.R
\docType{data}
\name{coagulation}
\alias{coagulation}
\title{Blood Coagulation Data on Four Distinct Diets}
\format{A data frame with 24 observations on each of the two variables.}
\usage{
coagulation
}
\description{
The effects of four different diets on a subject's blood coagulation time.
}
\details{
\itemize{
   \item \code{coag}: blood coagulation time.
   \item \code{diet}: Which diet they were on either A, B, C or D.
 }
}
\examples{
coagulation
}

