% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coag_descript.R
\docType{data}
\name{coagulation}
\alias{coagulation}
\title{Blood coagulation times by diet}
\format{a data frame with 24 rows and two columns:

\describe{
  \item{coag}{coagulation time in seconds}
  \item{diet}{diet type: A, B, C, or D}
}}
\source{
"Statistics for Experimenters" by G. P. Box, W. G. Hunter and J. S. Hunter, Wiley, 1978
}
\usage{
coagulation
}
\description{
Dataset comes from a study of blood coagulation times. 24 animals were
randomly assigned to four different diets and the samples were taken
in a random order.
}
\keyword{datasets}

