% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{Cohen_d_s}
\alias{Cohen_d_s}
\alias{Hedges_g_s}
\title{Effect size}
\usage{
Cohen_d_s(
  y1,
  y2,
  alternative = "two.sided",
  alpha = 0.05,
  mu = 0,
  return_CI = FALSE
)

Hedges_g_s(
  y1,
  y2,
  alternative = "two.sided",
  alpha = 0.05,
  mu = 0,
  return_CI = FALSE
)
}
\arguments{
\item{y1}{A numeric vector.}

\item{y2}{A numeric vector.}

\item{alternative}{Character (default: \code{"two.sided"}). Specifies the alternative hypothesis.
Available options are \code{c("two.sided", "less", "greater")}.}

\item{alpha}{Numeric (default: 0.05). Significance level (0 - 1) for hypothesis testing.}

\item{mu}{Numeric (default: 0).}

\item{return_CI}{Logical (default: FALSE). Whether to return the confidence interval of the effect size.}
}
\value{
A numeric scalar or a numeric vector of length 3, depends on \code{return_CI}.
}
\description{
Calculate Cohen's d and Hedges' g introduced in Lakens (2013).
}
\details{
Refer to Lakens (2013), the Cohen's d is the formula 1, the Hedges' g is the formula 4.
Their confidence interval (CI) were calculated based on formula 2, by transforming the \code{d} and \code{g}
back to \code{t} to fit the general confidence interval calculation.
For detailed and precise CI estimates, please use the \code{effectsize} package.
}
\references{
Lakens, D. (2013).
Calculating and reporting effect sizes to facilitate cumulative science: A practical primer for t-tests and ANOVAs.
Frontiers in Psychology, 4.
https://doi.org/10.3389/fpsyg.2013.00863
}
