% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_letter_display.R
\name{compact_letter_display}
\alias{compact_letter_display}
\title{Compact Letter Display (CLD)}
\usage{
compact_letter_display(
  x1,
  x2,
  pvalues,
  grp_names,
  centers,
  alpha = 0.05,
  descending = TRUE,
  display_letters = base::letters,
  display_null_letter = ""
)
}
\arguments{
\item{x1}{Character vector. The names of the minuend.}

\item{x2}{Character vector. The names of the subtrahend.}

\item{pvalues}{Numeric vector. The p-values for the differences of each \code{x1 - x2}.}

\item{grp_names}{Character vector. The group names (each factor levels).}

\item{centers}{Numeric vector. Generally, the corresponding mean or median values of \code{grp_names}.}

\item{alpha}{Numeric (default: 0.05). Significance level, range from 0 to 1.}

\item{descending}{Logical (default: TRUE). If \code{TRUE}, sort the centers in decreasing order.}

\item{display_letters}{Character vector (default: \code{base::letters}). Display symbols.}

\item{display_null_letter}{Character (default: ""). Symbol for filling the letters' gap.}
}
\value{
A character vector.
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\examples{
utils::data("iris", package = "datasets")
avg <- tapply(iris$Sepal.Length, iris$Species, "mean")
aov_mod <- stats::aov(Sepal.Length ~ Species, iris)
res <- stats::TukeyHSD(aov_mod)
res <- as.data.frame(res$Species)
res[["x1"]] <- unlist(lapply(strsplit(rownames(res), "-"), function(x) x[[1]]))
res[["x2"]] <- unlist(lapply(strsplit(rownames(res), "-"), function(x) x[[2]]))
compact_letter_display(
    x1 = res$x1,
    x2 = res$x2,
    pvalues = res$`p adj`,
    grp_names = names(avg),
    centers = unname(avg)
)
}
\references{
Piepho, H.-P. (2004).
An algorithm for a letter-based representation of all-pairwise comparisons.
Journal of Computational and Graphical Statistics, 13(2), 456–466.
https://doi.org/10.1198/1061860043515
Piepho, H.-P. (2018).
Letters in mean comparisons: What they do and don't mean.
Agronomy Journal, 110(2), 431–434.
https://doi.org/10.2134/agronj2017.10.0580
}
