% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway.factor}
\alias{oneway.factor}
\title{oneway.factor}
\usage{
\method{oneway}{factor}(z, y, ...)
}
\arguments{
\item{z}{A factor of levels for each observation}

\item{y}{A vector of responses}

\item{...}{Addition arguments used by other S3 methods}
}
\description{
S3 method for \code{\link{oneway}} using a response vector and factor
}
\seealso{
\code{link{oneway}}
}

