% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway.formula}
\alias{oneway.formula}
\title{oneway.formula}
\usage{
\method{oneway}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{A formula of the form \code{response~factor}}

\item{data}{An (optional) data frame used by the formula.}

\item{...}{Addition arguments used by other S3 methods}
}
\description{
An S3 method for \code{\link{oneway}} for formulas.
}
\seealso{
\code{\link{oneway}}
}

