% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oneway.R
\docType{package}
\name{oneway}
\alias{oneway}
\alias{oneway-package}
\alias{oneway.default}
\alias{oneway.formula}
\title{Oneway: One-way Analysis of Variance}
\usage{
oneway(z, ...)

\method{oneway}{default}(z, ...)

\method{oneway}{formula}(formula, data = list(), ...)
}
\arguments{
\item{z}{a data frame for the model.}

\item{formula}{a symbolic description of the model to be analyzed.}

\item{data}{an optional data frame containing the variables in the model.}

\item{\dots}{not used.}
}
\value{
An objet of class \code{ANOVA}, basically a list including elements:
\item{diet}{ diet }

\item{residuals}{ residuals }
}
\description{
The linmod package allow models to specified by their design matrix or by a formula.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: oneway.default(z, ...)

\item \code{formula}: oneway.formula(formula, data = list(), ...)
}}
\examples{
oneway(coagulation)
oneway(coag ~ diet, data = coagulation)
summary.oneway(coagulation)
}

