% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{Perform One-way ANOVA tests}
\usage{
oneway(formula, data)
}
\arguments{
\item{formula}{formula specifying dependent variable y and independent
variables x, written as "y ~ x"}

\item{data}{dataset which contains the variables specified in formula}
}
\value{
a list containing \textit{fit}, the model object,
and \textit{summarystats}, a dataframe of summary stats.
}
\description{
The oneway function performs a one way analysis of variance given a
formula and a data set.
}
\details{
The \code{oneway} function is essentially a wrapper for the \code{lm}
function, but returns instead a list containing the \code{lm} object
and a dataframe of summary stats. Additionally, plot, print, and summary
methods are provided under the class \textit{oneway}
}
\examples{
\dontrun{
# test for differences in mpg by number of cylinders (cyl) in mtcars
my_anova <- oneway(formula = mpg ~ as.factor(cyl), data = mtcars)
}
}
\seealso{
\link[stats]{lm}
}
