% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{One Way Analysis of Variance}
\usage{
oneway(formula, data)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{forumula}{an object of class formula, relating the dependent variable to the grouping
variable}
}
\value{
a list with 2 elements:
\item{oneway}{a list with lm results}
\item{summarystats}{a data frame with summary statistics}
}
\description{
\code{oneway} computes a oneway analysis of variance and inclues group
level summary statistics.
}
\details{
This function computes a standard oneway ANOVA, means, and standard deviations.
Missing values are handled via list-wise deletion
}
\author{
Shota Nakamura <snakamura@wesleyan.edu>
}
