% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oneway.R
\name{oneway}
\alias{oneway}
\title{One Way Analysis of Variance}
\usage{
oneway(x)
}
\arguments{
\item{x}{forumula an object of class formula, relating the dependent
variable to the grouping variable}

\item{data}{a data frame contain the variables in the model.}
}
\value{
a list with 2 elements:
\item{oneway}{a list with the lm result}
\item{summarystats}{a data frame with the summary statistics}
}
\description{
\code{oneway} computes a one-way analysis of variance
and inclues group level summary statistics
}
\details{
This function computes a standard one-way ANOVA, means,
and standard deviations. Missing values are handled via
listwise deletion.
}
\examples{
# mileage <- oneway(hwq~class, car)
summary(mileage)
mileage
plot(mileage)
}
\author{
Cher Qin \href{mailto:qqin@wesleyan.edu}{qqin@wesleyan.edu}
}
