% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{Perform a oneway analysis of variance}
\usage{
oneway(z, ...)
}
\arguments{
\item{z}{A list of responses grouped by factor level}

\item{...}{Addition arguments used by other S3 methods}
}
\description{
Perform a oneway analysis of variance
}
\examples{
library(faraway)
data(coagulation)
attach(coagulation)
oneway(split(coag, diet))
oneway(diet, coag)
oneway(coag~diet)
}
\seealso{
\code{\link{oneway.factor}}, \code{\link{oneway.formula}},
\code{\link{summary.oneway}}, \code{\link{plot.oneway}}, \code{\link{lsmeans}}
}

