% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{oneway_art}
\alias{oneway_art}
\title{One-Way Aligned Ranked Transformed Analysis of Variance (ART-ANOVA)}
\usage{
oneway_art(data, formula, alpha = 0.05, var_equal = NA, factor_levels = NULL)
}
\arguments{
\item{data}{A data frame containing the response and grouping variables.}

\item{formula}{A two-sided formula specifying the response and grouping
variables in the form \code{response ~ group}.}

\item{alpha}{A numeric significance level used to compute the critical
F-value. The default is \code{0.05}.}

\item{var_equal}{A logical value indicating whether equal variances should be
assumed. If \code{TRUE}, Fisher's ANOVA is performed. If \code{FALSE}, Welch's ANOVA
is performed. If \code{NA} (default), equality of variances is determined
automatically using \code{\link[varequal:is_var_equal]{varequal::is_var_equal()}}.}

\item{factor_levels}{Specify the order of the factor levels (default: NULL).}
}
\value{
A data frame representing the ANOVA table with the following columns:
\describe{
\item{DF}{Degrees of freedom.}
\item{SS}{Sum of squares.}
\item{MS}{Mean square.}
\item{Fvalue}{Observed F statistic.}
\item{Fcrit}{Critical F value at the specified significance level.}
\item{Pvalue}{P-value associated with the F statistic.}
\item{signif}{Significance code corresponding to the p-value.}
\item{method}{A character specifying this is an ART-ANOVA procedure.}
}

The rows correspond to the treatment groups ("Group"), residual error
("Residuals"), and total variation ("Total").
}
\description{
Performs a one-way analysis of variance (ANOVA) to compare the means of two
or more independent groups. By default, the function automatically selects
between the classical Fisher ANOVA and Welch's ANOVA according to whether
the group variances are judged to be homogeneous.
}
\details{
If \code{var_equal = TRUE}, Fisher's ANOVA is performed on the transformed response variable.
If \code{var_equal = FALSE}, Welch's ANOVA is performed. When \code{var_equal = NA} (the default),
homogeneity of variances is assessed using \code{\link[varequal:is_var_equal]{varequal::is_var_equal()}}, and the appropriate
test is selected automatically.
}
\examples{
oneway_art(anorexia, weight_gain ~ therapy)

}
\references{
Wobbrock, J. O., Findlater, L., Gergle, D., & Higgins, J. J. (2011).
The aligned rank transform for nonparametric factorial analyses using only ANOVA procedures.
Proceedings of the SIGCHI Conference on Human Factors in Computing Systems, 2011, 143–146.
https://doi.org/10.1145/1978942.1978963

Elkin, L. A., Kay, M., Higgins, J. J., & Wobbrock, J. O. (2021).
An aligned rank transform procedure for multifactor contrast tests.
Proceedings of the 34th Annual ACM Symposium on User Interface Software and Technology, 754–768.
https://doi.org/10.1145/3472749.3474784
}
