% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot one-way ANOVA}
\usage{
\method{plot}{oneway}(x, col = "skyblue", ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{col}{fill color for boxplots}

\item{...}{additional arguments passed to the function passed to the
\code{\link{boxplot}} function.}
}
\description{
\code{plot.oneway} group comparisons for a one-way ANOVA.
}
\details{
This function creates side by side box plots for each group
in the analysis.
}
\examples{
mileage <- oneway(hwy ~ class, cars)
plot(mileage)
}
\author{
Rob Kabacoff <rkabacoff@wesleyan.edu>
}
