% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oneway.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot method for oneway.}
\usage{
\method{plot}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"oneway"} i.e., an analysis of variance.}

\item{\dots}{not used.}
}
\description{
Plot method for oneway.
}

