% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot function for Oneway objects}
\usage{
\method{plot}{oneway}(x, col = "skyblue", ...)
}
\arguments{
\item{x}{Object of class "oneway"}

\item{...}{parameters passed to boxplot function}
}
\value{
a list
}
\description{
Produces a boxplot which follows the formula specified in the oneway object
}
\examples{
\dontrun{
aov1 <- oneway(mpg ~ cyl, mtcars)
plot(aov1)
}
}
\seealso{
\link[graphics]{boxplot}
}
