% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot one-way ANOVA}
\usage{
\method{plot}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class one way \code{oneway}}

\item{...}{additional arguments passed to the \code{\link{boxplot}} function}

\item{col}{Fill color for boxplots}
}
\description{
\code{plot.oneway} Plots a oneway anova
}
\details{
This function creates side by side boxplots for each group in analysis
}
\author{
Shota Nakamura <snakamura@wesleyan.edu>
}
