% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Creating boxplot of groups in a \code{oneway} object}
\usage{
\method{plot}{oneway}(x, names = x$groups, xlab = "Grouping variable",
  ylab = "Response variable", main = capture.output(x$call))
}
\arguments{
\item{x}{A \code{oneway} object}

\item{names}{Names of factor levels}

\item{xlab}{X label}

\item{ylab}{Y label}

\item{main}{Main plot title}

\item{...}{Optional graphing arguments to be passed to
\code{boxplot}}
}
\description{
Creating boxplot of groups in a \code{oneway} object
}
\seealso{
\code{link{boxplot}}
}

