% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot one-way ANOVA}
\usage{
\method{plot}{oneway}(x, ...)
}
\arguments{
\item{x}{is an object of class oneway.}

\item{...}{additional argument passed to boxplot function.}
}
\description{
plot.oneway create groups comparison for a one-way ANOVA
}
\examples{
# mileage <- oneway(mpg~cyl, mtcars)
plot(mileage)
}
