% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot one-way ANOVA}
\usage{
\method{plot}{oneway}(x, plot = "boxplot", ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{plot}{a character string representing the type of plot. Default is \code{boxplot}.
Other types of plot: \code{violin} and \code{density_ridges}.}

\item{...}{additional arguments passed to the chosen plot function.}
}
\value{
a ggplot2 graph
}
\description{
\code{plot.oneway} creates group comparisons for a one-way ANOVA and gives the user an option to choose among several types of graphs.
}
\examples{
fit <- oneway(mpg ~ cyl, mtcars)
plot(fit, plot = "boxplot")
plot(fit, plot = "violin")
plot(fit, plot = "density_ridges")
}
