% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway-plot.r
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot one-way ANOVA}
\usage{
\method{plot}{oneway}(x, plot = "boxplot", ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{plot}{a character string representing the type of plot. Default is boxplot. Other types of plot: violin and density ridges.}

\item{...}{additional arguments passed to the chosen plot function.}
}
\description{
\code{plot.oneway} creates group comparisons for a one-way ANOVA and gives the user an option to choose among several types of graphs.
}
\examples{
fit <- oneway(mpg ~ cyl, mtcars)
plot(fit, plot = "boxplot")
plot(fit, plot = "violin")
plot(fit, plot = "density_ridges")
}
