% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot one-way ANOVA}
\usage{
\method{plot}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{...}{additional arguments passed to the
\code{\link{boxplot}} function.}

\item{col}{fill color for boxplots}

\item{data}{a data frame containing the variables 
in the model.}
}
\description{
\code{plot.oneway} plots a one-way anova
}
\details{
This function creates side by side box plots for each group
in the analysis.
}
\examples{
mileage <- oneway(hwy ~ class, cars)
plot(mileage)
}
\author{
Shane Ross <saross@wesleyan.edu>
}
