% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Print one-way ANOVA results}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{...}{additional arguments passed to the function.}
}
\value{
the input object is returned silently
}
\description{
\code{print.oneway} prints one-way ANOVA results
}
\details{
This function prints one-way ANOVA results created
by the \code{\link{oneway}} function.
}
\examples{
mileage <- oneway(hwy ~ class, cars)
plot(mileage)

}
\author{
Rob Kabacoff <rkabacoff@wesleyan.edu>
}
