% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Print one-way ANOVA results}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class oneway.}

\item{...}{additional arguments passed to the function.}
}
\value{
the input object is returned silently
}
\description{
print.one-ways prints one way anova results
}
\examples{
mileage <- oneway(mpg ~ cyl, mtcars)
print(mileage)
}
