% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oneway.R
\name{print.oneway}
\alias{print.oneway}
\title{Default print for oneway.}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"oneway"}, i.e., a an analysis of variance.}

\item{\dots}{not used.}
}
\description{
Default print for oneway.
}

