% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Print function for oneway objects}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{object of class "oneway"}

\item{...}{parameters passed to print function}
}
\description{
Print function for oneway objects
}
\examples{
\dontrun{
aov1 <- oneway(mpg ~ cyl, mtcars)
print.oneway(aov1)
}
}
\seealso{
\link[base]{print}
}
