% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Print one-way ANOVA}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class one way \code{oneway}}

\item{...}{additional arguments passed to the \code{\link{print}} function}
}
\value{
input object is returned silently (x gets returned)
}
\description{
\code{print.oneway} Prints a oneway anova
}
\details{
This function prints oneway ANOVA results created by the oneway function
}
\author{
Shota Nakamura <snakamura@wesleyan.edu>
}
