% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Print one-way ANOVA}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{...}{additional arguments passed to the
\code{\link{boxplot}} function.}

\item{col}{fill color for boxplots}

\item{data}{a data frame containing the variables 
in the model.}
}
\value{
the input object is returned silently
}
\description{
\code{print.oneway} print a one-way anova
}
\details{
This function creates prints one-way ANOVA results created by
the \code{\link{oneway}} function
}
\examples{
mileage <- oneway(hwy ~ class, cars)
mileage
}
\author{
Shane Ross <saross@wesleyan.edu>
}
