% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oneway.R
\name{print.summary.oneway}
\alias{print.summary.oneway}
\title{Print method for the summary method.}
\usage{
\method{print}{summary.oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"summary.oneway"}, i.e., an analysis of variance.}

\item{\dots}{not used.}
}
\description{
Print method for the summary method.
}

