% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summarize One-Way ANOVA Results}
\usage{
\method{summary}{oneway}(x)
}
\arguments{
\item{x}{an onject of class 'oneway'.}
}
\value{
object of type anova and data.frame
}
\description{
summary.oneway summarizes the results of the one-way ANOVA
}
\examples{
# mileage <- oneway(mpg ~ cyl, mtcars)
summary(mileage)
}
