% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oneway.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summary method for oneway.}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{an objet of class \code{"oneway"}, an analysis of variance.}

\item{\dots}{not used.}
}
\description{
Summary method for oneway.
}

