% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summarizing oneway class ANOVA models}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of type oneway}

\item{...}{parameters passed to print}
}
\value{

}
\description{
produces summary table of the anova model's results
}
\examples{
\dontrun{
aov1 <- oneway(mpg ~ cyl, mtcars)
print.oneway(aov1)
}
}
\seealso{
\link[base]{print}
}
