% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summary one-way ANOVA}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class one way \code{oneway}}

\item{...}{additional arguments passed to the \code{\link{print}} function}
}
\value{
input object is returned silently (x gets returned)
}
\description{
\code{summary.oneway} Summarize a oneway ANOVA results
}
\details{
This function summarises the oneway ANOVA results created by the oneway function
}
\author{
Shota Nakamura <snakamura@wesleyan.edu>
}
