% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{summary.oneway}
\alias{summary.oneway}
\title{summary.oneway}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{An object of class \code{oneway}}

\item{...}{Addition arguments used by other S3 methods}
}
\description{
Creates an Analysis of Variance table for a \code{oneway} object
}

