% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summarize one-way ANOVA}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{an object of class oneway.}

\item{...}{additional options. not currently used.}
}
\value{
object of type anova and data.frame
}
\description{
summary.oneway summarizes the results of the one-way ANOVA
}
\examples{
fit <- oneway(mpg ~ cyl, mtcars)
summary(fit)
}
