% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Plot one-way ANOVA}
\usage{
\method{summary}{oneway}(x)
}
\arguments{
\item{x}{an object of class oneway.}
}
\value{
object of type anova and dataframe
}
\description{
summary.oneway summarizes the results of the one-way ANOVA
}
\examples{
mileage <- oneway(mpg ~ cyl, mtcars)
summary(mileage)
}
