% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summarize one-way ANOVA results}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{...}{additional arguments passed to the print function}

\item{data}{a data frame containing the variables 
in the model.}
}
\value{
the input object is returned silently
}
\description{
\code{summarize.oneway} summarizes the results of
  a one-way anova
}
\details{
This function prints the ANOVA table from the linear model
}
\examples{
mileage <- oneway(hwy ~ class, cars)
summary(mileage)
}
\author{
Shane Ross <saross@wesleyan.edu>
}
